requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
wikipedia>=1.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
openai>=1.0.0
huggingface-hub>=0.19.0
datasets>=2.16.0
rouge-score>=0.1.2
tqdm>=4.66.0
pydantic>=2.5.0
python-dotenv>=1.0.0
lxml>=4.9.0